+++ src/main/java/com/simsilica/sim/GameLoop.java
@@ -47,18 +47,27 @@ import java.util.concurrent.atomic.AtomicBoolean;
  *  @author    Paul Speed
  */
 public class GameLoop {
+
+    private static final long INTERVAL_FPS_60 = 16666667L; 
+
     private final GameSystemManager systems;
     private final Runner loop = new Runner(); 
  
     private long updateRate;
+    private long idleSleepTime = 0;
     
     public GameLoop(GameSystemManager systems) {
-        this(systems, 16666667L); // 60 FPS 
+        this(systems, INTERVAL_FPS_60); // 60 FPS 
     }
     
     public GameLoop( GameSystemManager systems, long updateRateNanos ) {
+        this(systems, updateRateNanos, null);
+    }
+    
+    public GameLoop( GameSystemManager systems, long updateRateNanos, Long idleSleepTime ) {
         this.systems = systems;
         this.updateRate = updateRateNanos;
+        setIdleSleepTime(idleSleepTime);
     }   
 
     public GameSystemManager getGameSystemManager() {
@@ -84,6 +93,35 @@ public class GameLoop {
      */   
     public void stop() {
         loop.close();
+    }
+ 
+    /**
+     *  Sets the period of time in milliseconds that the update loop will
+     *  wait between time interval checks.  This defaults to 0 because on
+     *  Windows the default 60 FPS update rate will cause frame drops for
+     *  a idle sleep time of 1 or more.  However, 0 causes the loop to consume
+     *  a noticable amount of CPU.  For lower framerates, 1 is recommended
+     *  and will be set automically based on the current frame rate interval
+     *  if null is specified.
+     */
+    public final void setIdleSleepTime( Long millis ) {
+        if( millis == null ) {
+            // Configure a reasonable default
+            if( updateRate > INTERVAL_FPS_60 ) {
+                // Can probably get away with more sleep
+                this.idleSleepTime = 1;
+            } else {
+                // Else on Windows, sleep > 0 will take almost as long as
+                // a frame
+                this.idleSleepTime = 0;
+            }
+        } else {
+            this.idleSleepTime = millis;
+        }
+    }
+
+    public Long getIdleSleepTime() {
+        return idleSleepTime;
     }
  
     /**
@@ -121,18 +159,41 @@ public class GameLoop {
             long lastTime = System.nanoTime();
             while( go.get() ) {
                 long time = System.nanoTime();
-                long delta = time - lastTime; 
-                if( delta < updateRate ) {
-                    // Not time to update yet
+                long delta = time - lastTime;
+                 
+                if( delta >= updateRate ) {
+                    // Time to update
+                    lastTime = time;                                        
+                    systems.update();
                     continue;
                 }
-                lastTime = time;                                        
-                systems.update();
-                
+ 
                 // Wait just a little.  This is an important enough thread
-                // that we'll poll instead of smart-sleep. 
+                // that we'll poll instead of smart-sleep but we also don't
+                // want to consume 100% of a CPU core. 
+                // Thread.sleep(0) relies on the operating systems thread
+                // scheduler which will have a minimum granularity.  On Windows, 
+                // Thread.sleep(0) can take as long as 15 ms to return.  Thus
+                // note that if we devolve into a case where the delta is more
+                // than the update rate then we never hit this code.
+                //
+                // And if the delta since last poll was kind of large
+                // we won't sleep at all... In this case, if the delta was
+                // more than half the updateRate... we'll skip sleeping.
+                // Which probably means that on Windows, we'll be responsive
+                // every other loop iteration.  It's an imperfect world and
+                // the only alternative is true CPU heating busy-waiting.
+                if( delta * 2 > updateRate ) {
+                    continue;
+                } 
                 try {
-                    Thread.sleep(0);
+                    // More then 0 here and we underflow constantly.
+                    // 0 and we gobble up noticable CPU.  Slower update
+                    // rates could probably get away with the longer sleep
+                    // but 60 FPS can't keep up on Windows.  So we'll let
+                    // the sleep time be externally defined for cases
+                    // where the caller knows  better.
+                    Thread.sleep(idleSleepTime);
                 } catch( InterruptedException e ) {
                     throw new RuntimeException("Interrupted sleeping", e);
                 }