+++ release-notes.md
@@ -1,4 +1,12 @@
 
+Version 1.0.3
+--------------
+* Flipped the GameLoop's update loop to sleep when idle instead of
+    only after an actual update was run.
+* Modified GameLoop to allow configuring the amount of time to sleep
+    when idle-busy-polling for the next update interval.
+    
+
 Version 1.0.2
 --------------
 * Added some lifecycle trace logging to GameSystemManager.
@@ -8,6 +16,7 @@ Version 1.0.2
     of style settings.
 * Fixed a bug where the wrong class was being reported in the exception message for
     adding event bus listeners.     
+
 
 Version 1.0.1
 --------------